requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
python-multipart>=0.0.20,<1.0
openpyxl>=3.1,<4.0
pydantic>=2.11,<3.0
python-docx>=1.1,<2.0
reportlab>=4.4,<5.0
openai>=1.0,<2.0
httpx>=0.28,<1.0
beautifulsoup4>=4.12,<5.0
argon2-cffi>=23.1,<26.0
cryptography>=44.0,<46.0
pypdf>=6.9,<7.0
SQLAlchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
